Extract template generation to seperate file
