+++ src/smt/mod.rs
@@ -18,7 +18,6 @@ use z3rro::{
 
 pub mod funcs;
 
-pub mod inv_synth_helpers;
 pub mod pretty_model;
 pub mod symbolic;
 mod symbols;
